chore: provide context for event instead of passing props
